ref(cmdk): Restore CommandPaletteHotkeys component and remove duplicate hotkey registration
Move cmd+k hotkey logic back into CommandPaletteHotkeys (now using
useGlobalModal instead of the deleted state.open field) and render it
from navigation/index.tsx, removing the duplicate inline useHotkeys call.
