+++ static/app/views/navigation/index.tsx
@@ -4,10 +4,7 @@ import styled from '@emotion/styled';
 import {Container, Flex} from '@sentry/scraps/layout';
 import {ExternalLink} from '@sentry/scraps/link';
 
-import {
-  openCommandPaletteDeprecated,
-  toggleCommandPalette,
-} from 'sentry/actionCreators/modal';
+import {CommandPaletteHotkeys} from 'sentry/components/commandPalette/ui/commandPaletteStateContext';
 import {useGlobalCommandPaletteActions} from 'sentry/components/commandPalette/useGlobalCommandPaletteActions';
 import {useGlobalModal} from 'sentry/components/globalModal/useGlobalModal';
 import {t} from 'sentry/locale';
@@ -33,7 +30,6 @@ import {useHasPageFrameFeature} from 'sentry/views/navigation/useHasPageFrameFea
 import {useResetActiveNavigationGroup} from 'sentry/views/navigation/useResetActiveNavigationGroup';
 
 function UserAndOrganizationNavigation() {
-  const organization = useOrganization();
   const {layout} = usePrimaryNavigation();
   const {visible} = useGlobalModal();
   const {view, setView} = useSecondaryNavigation();
@@ -52,22 +48,9 @@ function UserAndOrganizationNavigation() {
         ]
   );
 
-  useHotkeys([
-    {
-      match: ['command+shift+p', 'command+k', 'ctrl+shift+p', 'ctrl+k'],
-      includeInputs: true,
-      callback: () => {
-        if (organization.features.includes('cmd-k-supercharged')) {
-          toggleCommandPalette({}, organization, visible, 'keyboard');
-        } else {
-          openCommandPaletteDeprecated();
-        }
-      },
-    },
-  ]);
-
   return (
     <NavigationLayout>
+      <CommandPaletteHotkeys />
       {layout === 'mobile' ? (
         <MobileSecondaryNavigationContextProvider>
           {hasPageFrame ? <MobilePageFrameNavigation /> : <MobileNavigation />}
